+++ interface/database/pg/update_1.1.0.sql
@@ -1,0 +1,2 @@
+-- add index on geocube.records on datetime
+CREATE INDEX idx_records_datetime ON geocube.records (datetime);